dom/xslt/tests/mochitest/test_bug1072116.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1072116
-->
<head>
  <title>Test for Bug 1072116</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1072116">Mozilla Bug 1072116</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 1072116 **/
var attr = document.createAttribute("c");

var xpathExpr = document.createExpression('a', null);

var status = false;
try {
  xpathExpr.evaluate(attr, null, null, null, null);
} catch(e) {
  status = true;
}

ok(status, "Still alive \\o/");

</script>
</pre>
</body>
</html>
